requirements.txt
requests
Pillow
flask 